pyproject.toml
[project]
name = "wk"
version = "0.1.0"
description = "网课学习助手"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.0.0",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
]
